requirements.txt
Flask 
Flask-SQLAlchemy
Flask-Bcrypt
PyJWT
flask-cors